+++ ml-dsa/Cargo.toml
@@ -39,7 +39,7 @@ signature = { version = "3.0.0-rc.4", default-features = false, features = ["dig
 
 # optional dependencies
 const-oid = { version = "0.10", features = ["db"], optional = true }
-pkcs8 = { version = "0.11.0-rc.6", default-features = false, optional = true }
+pkcs8 = { version = "0.11.0-rc.7", default-features = false, optional = true }
 rand_core = { version = "0.9", optional = true }
 zeroize = { version = "1.8.1", optional = true, default-features = false }
