--- ml-dsa/tests/examples/ML-DSA-44.priv
@@ -1,4 +1,0 @@
------BEGIN PRIVATE KEY-----
-MDICAQAwCwYJYIZIAWUDBAMRBCAAAQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRob
-HB0eHw==
------END PRIVATE KEY-----
