--- ml-dsa/tests/examples/ML-DSA-65.priv
@@ -1,4 +1,0 @@
------BEGIN PRIVATE KEY-----
-MDICAQAwCwYJYIZIAWUDBAMSBCAAAQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRob
-HB0eHw==
------END PRIVATE KEY-----
