--- ml-dsa/tests/examples/ML-DSA-87.priv
@@ -1,4 +1,0 @@
------BEGIN PRIVATE KEY-----
-MDICAQAwCwYJYIZIAWUDBAMTBCAAAQIDBAUGBwgJCgsMDQ4PEBESExQVFhcYGRob
-HB0eHw==
------END PRIVATE KEY-----
